Create a separate class for StoreTransaction
